+++ perun/utils/helpers.py
@@ -34,6 +34,7 @@ IndexEntry = collections.namedtuple("IndexEntry", "time checksum path offset")
 
 # Minor Version specific things
 MinorVersion = collections.namedtuple("MinorVersion", "date author email checksum desc parents")
+MajorVersion = collections.namedtuple("MajorVersion", "name head")
 
 # Profile specific stuff
 SUPPORTED_PROFILE_TYPES = ['memory', 'mixed', 'time']